+++ vm/sandbox_primitives.go
@@ -52,6 +52,7 @@ func (vm *VM) registerSandboxPrimitives() {
 			defer func() {
 				v.HandleForkedPanic(proc, recover())
 				v.unregisterInterpreter()
+				v.registry.ReleaseBlock(blockVal)
 			}()
 
 			interp := v.newForkedInterpreter(hidden)